- fixing GCC warnings (0001-Compile-Fix-gcc-warnings.patch).

+++ src/jscrypt/cipher.cpp
@@ -244,7 +244,7 @@ DEFINE_CONSTRUCTOR() {
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
 			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
-			JL_S_ASSERT( optargLength == (int)keyLength, "The tweak length must be %d bytes length (key size)", (int)keyLength );
+			JL_S_ASSERT( optargLength == keyLength, "The tweak length must be %d bytes length (key size)", (int)keyLength );
 			err = lrw_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (signed)keyLength, (unsigned char *)optarg, numRounds, (symmetric_LRW *)pv->symmetric_XXX );
 			break;
 		}
@@ -307,7 +307,7 @@ DEFINE_FUNCTION( Encrypt ) {
 	int err;
 	switch ( pv->mode ) {
 		case mode_ecb:
-			JL_S_ASSERT( ptLength == pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
+			JL_S_ASSERT( ptLength == (size_t)pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
 			err = ecb_encrypt( (const unsigned char *)pt, (unsigned char *)ct, ptLength, (symmetric_ECB *)pv->symmetric_XXX );
 			break;
 		case mode_cfb:
@@ -317,14 +317,14 @@ DEFINE_FUNCTION( Encrypt ) {
 			err = ofb_encrypt( (const unsigned char *)pt, (unsigned char *)ct, ptLength, (symmetric_OFB *)pv->symmetric_XXX );
 			break;
 		case mode_cbc:
-			JL_S_ASSERT( ptLength == pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
+			JL_S_ASSERT( ptLength == (size_t)pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
 			err = cbc_encrypt( (const unsigned char *)pt, (unsigned char *)ct, ptLength, (symmetric_CBC *)pv->symmetric_XXX );
 			break;
 		case mode_ctr:
 			err = ctr_encrypt( (const unsigned char *)pt, (unsigned char *)ct, ptLength, (symmetric_CTR *)pv->symmetric_XXX );
 			break;
 		case mode_lrw:
-			JL_S_ASSERT( ptLength == pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
+			JL_S_ASSERT( ptLength == (size_t)pv->descriptor->block_length, "This mode require a %d bytes block of data", pv->descriptor->block_length );
 			err = lrw_encrypt( (const unsigned char *)pt, (unsigned char *)ct, ptLength, (symmetric_LRW *)pv->symmetric_XXX );
 			break;
 		case mode_f8:
@@ -480,7 +480,7 @@ DEFINE_PROPERTY( IVSetter ) {
 			JL_REPORT_ERROR("No IV in ECB mode");
 		case mode_cfb: {
 			symmetric_CFB *tmp = (symmetric_CFB *)pv->symmetric_XXX;
-			JL_S_ASSERT( IVLength == tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
+			JL_S_ASSERT( IVLength == (size_t)tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
 			err = cfb_setiv( (const unsigned char *)IV, IVLength, tmp );
 			if (err != CRYPT_OK)
 				return ThrowCryptError(cx, err);
@@ -488,7 +488,7 @@ DEFINE_PROPERTY( IVSetter ) {
 		}
 		case mode_ofb: {
 			symmetric_OFB *tmp = (symmetric_OFB *)pv->symmetric_XXX;
-			JL_S_ASSERT( IVLength == tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
+			JL_S_ASSERT( IVLength == (size_t)tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
 			err = ofb_setiv( (const unsigned char *)IV, IVLength, tmp );
 			if (err != CRYPT_OK)
 				return ThrowCryptError(cx, err);
@@ -496,7 +496,7 @@ DEFINE_PROPERTY( IVSetter ) {
 		}
 		case mode_cbc: {
 			symmetric_CBC *tmp = (symmetric_CBC *)pv->symmetric_XXX;
-			JL_S_ASSERT( IVLength == tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
+			JL_S_ASSERT( IVLength == (size_t)tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
 			err = cbc_setiv( (const unsigned char *)IV, IVLength, tmp );
 			if (err != CRYPT_OK)
 				return ThrowCryptError(cx, err);
@@ -504,7 +504,7 @@ DEFINE_PROPERTY( IVSetter ) {
 		}
 		case mode_ctr: {
 			symmetric_CTR *tmp = (symmetric_CTR *)pv->symmetric_XXX;
-			JL_S_ASSERT( IVLength == tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
+			JL_S_ASSERT( IVLength == (size_t)tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
 			err = ctr_setiv( (const unsigned char *)IV, IVLength, tmp );
 			if (err != CRYPT_OK)
 				return ThrowCryptError(cx, err);
@@ -520,7 +520,7 @@ DEFINE_PROPERTY( IVSetter ) {
 		}
 		case mode_f8: {
 			symmetric_F8 *tmp = (symmetric_F8 *)pv->symmetric_XXX;
-			JL_S_ASSERT( IVLength == tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
+			JL_S_ASSERT( IVLength == (size_t)tmp->blocklen, "This cipher require a IV length of %d", tmp->blocklen );
 			err = f8_setiv( (const unsigned char *)IV, IVLength, tmp );
 			if (err != CRYPT_OK)
 				return ThrowCryptError(cx, err);
